// src/app/models/registration/Registration.model.ts
export class Registration {
  //id: number;
  //name: string;
  //password: string;
  //roleId: number;
  //DateOfJoining: Date;
  //Designation: string;
  //Phone: number;
  //DateOfBirth: Date;

  //AddressDescription1: string;
  //AddressDescription2: string;

  //City: string;
  //State: string;
  //ZipCode: number;

  id: number;
  name: string;
  password: string;
  roleid: number;
  DateOfJoining: Date;
  designation: string;
  phone: number;
  DateOfBirth: Date;
  addressDescription1: string;
  addressDescription2: string;
  city: string;
  state: string;
  zipCode: number;
}